Update refresher to use ipc cert
