+++ lib/saluki-config/Cargo.toml
@@ -10,6 +10,7 @@ arc-swap = { workspace = true }
 figment = { workspace = true, features = ["env"] }
 reqwest = { workspace = true, features = ["rustls-tls-native-roots-no-provider", "json"] }
 saluki-error = { workspace = true }
+saluki-io = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
